% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulateHelpers.R
\name{reticulate_conda_list_packages}
\alias{reticulate_conda_list_packages}
\title{List Conda Packages}
\usage{
reticulate_conda_list_packages(envname = NULL, conda = "auto", no_pip = TRUE)
}
\description{
This is almost a direct copy of the unexported
  reticulate::conda_list_packages function and the other functions needed to
  run it that are also unexported from {reticulate}. Need to look into how to
  apply the copyright notice for the Apache 2.0 license.
}
