% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulateHelpers.R
\name{reticulate_miniconda_exists}
\alias{reticulate_miniconda_exists}
\title{Does MiniConda Exist?}
\usage{
reticulate_miniconda_exists(path = reticulate::miniconda_path())
}
\description{
Does MiniConda Exist?
}
